add field and operation models
